pyproject.toml
[tool.poetry]
name = "storiescreator"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
feedparser = "^6.0.2"
lxml = "^4.6.2"
replit = "^2.0.0"
"pushbullet.py" = "^0.12.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
